Separate zeroshot-specific code
