+++ pytorch/torch/utils/test_utils.py
@@ -9,11 +9,7 @@ class TestDataLoader:
 		import pytorch.globals
 
 		from torchvision.transforms import CenterCrop, Compose, Resize
-		from pytorch.torchvision.datasets import (
-			AnimalsWithAttributesDataset,
-			ZeroshotAnimalsWithAttributesDataset,
-			TransductiveZeroshotAnimalsWithAttributesDataset,
-		)
+		from pytorch.torchvision.datasets import AnimalsWithAttributesDataset
 		from pytorch.torch.utils.data import AnimalsWithAttributesDataLoader
 
 	#	Transform to be used on images:
@@ -28,27 +24,9 @@ class TestDataLoader:
 	#	Batch size:
 		batch_size = 64
 
-	#	Plain dataloader:
+	#	Dataloader:
 		dataloader = AnimalsWithAttributesDataLoader(
 			AnimalsWithAttributesDataset(
-				transform=transform,
-			), pytorch.globals.generator, batch_size)
-		image, label = next(iter(dataloader))
-		assert image.size()[0] == batch_size
-		assert label.size()[0] == batch_size
-
-	#	Zeroshot dataloader:
-		dataloader = AnimalsWithAttributesDataLoader(
-			ZeroshotAnimalsWithAttributesDataset(
-				transform=transform,
-			), pytorch.globals.generator, batch_size)
-		image, label = next(iter(dataloader))
-		assert image.size()[0] == batch_size
-		assert label.size()[0] == batch_size
-
-	#	Transductive zeroshot dataloader:
-		dataloader = AnimalsWithAttributesDataLoader(
-			TransductiveZeroshotAnimalsWithAttributesDataset(
 				transform=transform,
 			), pytorch.globals.generator, batch_size)
 		image, label = next(iter(dataloader))
